v0.3.1020 - /drawings leaves client.ts
SCALE-SEAM ⑮. Eleven sheet methods (revise, schedules, storeys, sync-status,
markup + stream) move to api/drawingSheets.ts. Six regions.
client.ts 3,538 → 3,482.
