+++ apps/web/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@aec/web",
-    "version": "0.3.1019",
+    "version": "0.3.1020",
   "private": true,
   "type": "module",
   "engines": {
