+++ apps/web/src-tauri/tauri.conf.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://schema.tauri.app/config/2",
   "productName": "Massing",
-  "version": "0.3.1019",
+  "version": "0.3.1020",
   "identifier": "com.ibuilder.aecbim",
   "build": {
     "frontendDist": "../dist",
